requirements.txt
numpy
matplotlib
scipy
h5py
dask
cython
pillow
colorcet
pyopengl
requests
qtconsole
tqdm
joblib
click
